split test cases by file & create tests for useUser
